false start with another tree representation
I thought that representing our tree more like an algebraic data type
would help the code to be cleaner.  It cost a lot of code and still
wasn't clean.  I'm going to throw this away and try something
different, but wanted this for posterity in case I come back to it
later.