Only decrypt keys when needed
Prior to this commit, using "hush set" to modify one leaf caused the
entire file to be re-encrypted.  For hush files stored under version
control, that causes many irrelevant changes.

Now we keep leaves in encrypted form as long as possible.  We only
decrypt when printing.  We encrypt new values as soon as we store them
into the tree.

This produces a hush file whose contents are much more stable than
before.