Make Decrypt return a new tree
It made me nervous to decrypt the tree in place.  If someone
accidentally called tree.Print() and then tree.Save(), they'd write a
file with all the secrets exposed.

I'd actually like to encode the distinction between encrypted and
decrypted values in the type system somehow to reduce the chances of
that kind of mistake.