Safe permissions for hush file
Leaves in the hush file are encrypted, but the tree structure
(which is plaintext) may contain sensitive information, depending on
the user's naming conventions.  For example, it might contain the
names of websites that the user frequents.

By restricting read and write permissions to the user only, it gives
some assurance of privacy for the tree structure.

This highlights a privacy vs convenience trade off. If we were
optimizing privacy, we'd encrypt the entire hush file.  Unfortunately,
that would make diffs and using a regular text editor unpleasant.
Leaving the entire file unencrypted would be especially convenient to
work with, but risks divulging plaintext passwords if the file is
lost.