Rough start on "init" command
When the user first begins using hush, they must initialize the system
by selecting a password.  We use that opportunity to generate
encryption keys and establish a basic tree for them.

This sketch does the password and key steps, but doesn't create the
tree.  Once I got this far, I realized that tree handling is getting
too complex and I want to lean on the type system to help me keep
everything straight. I alluded to this wish in commit adcf5254160.
Now it's time to make it happen.