Make Value type keep track of encoding and privacy details
A value used to be just a string.  I realized that I was having a hard
time keeping track of all the different kinds of strings it might
be (base64 encoded ciphertext, base64 decoded private plaintext,
base64 decoded public plaintext, etc).

Now a value is a struct which keeps all of this straight for me.  If I
ask it to do something that doesn't make sense (like decrypt a
plaintext), it panics so that I can repair my broken assumptions.