Use password for encryption
The old code used a hardcoded encryption key for encrypting secrets.
With this commit, real encryption is in place.

The user provides a password.  That combined with a public salt gives
us a "password key".  We use the password key to decrypt the
"encryption key".  We use the encryption key to decrypt the tree's
leaves.

The extra level of indirection with a password key and a separate
encryption key are not for added security.  They're there for
convenience.  Eventually, the user will be able to change his password
without us having to rewrite the entire file.  We'll just change the
salt and reencrypt the encryption key.  All other leaves can stay as
they are.  When tracking the hush file in version control, this
restricts the diffs to only those pieces that truly changed.