Completely change the Tree representation
While working with the tree, I kept needing a sorted list of key-value
pairs.  I decided it would be easier to just keep the tree in memory
as a list of key-value pairs (with an index).  The Tree can implement
sort.Interface so that the sorted list is easily accessible.

This should make it possible to verify a tree's checksum.  I think
it'll also make several simplifications possible.