Store keys in initial tree
When initializing the hush tree, both encryption and MAC keys belong
in the tree.  Without keys in the tree, results are subtly wrong.

The new panics in this commit would have loudly flagged the mistake
earlier if they had been in place.  They should ensure that we don't
try to encrypt or HMAC anything without a key, which is obviously not
what we intend.