Add rm command
One can now remove portions of the hush file from the command line.

This commit suggests that we'll want to represent the tree differently
than we do right now.  It feels like we're accumulating too much code
for tree management and there should be a simpler, cleaner way to do
it.  I'm hesitant to implement a full balanced tree.  For now, I'll
keep accumulating cruft until I'm satisfied that I know how we're
using the tree.  Then we can choose a good data structure to represent
it.