Consistently ignore write errors
We ignore all write errors in Tree.Save() but one call captured the
value.  This made it look like we planned to handle this error but
forgot.  By not capturing the value, it's clearer that we intend to
ignore it.

Even clearer would have been to use _ for every Write's error value.
That felt like a lot of code motion for little gain.