Try to avoid rename failures
We used to create temporary files in the default location for
temporary files.  Unfortunately, that's often on a separate file
system from the hush file itself which increases the chances of the
rename system call failing.

By creating our temporary file in the same directory as the hush file,
it increases the chances that rename will succeed atomically and avoid
losing data during a failed copy.